Remove StringMatcher and integrate functionality in StringEqualityResult
