+++ Source/aweXpect.Core/Core/IOptionsEquality.cs
@@ -6,8 +6,8 @@
 public interface IOptionsEquality<in T>
 {
 	/// <summary>
-	///     Returns <see langword="true" /> if the two objects <paramref name="a" /> and <paramref name="b" /> are considered
+	///     Returns <see langword="true" /> if the two objects <paramref name="actual" /> and <paramref name="expected" /> are considered
 	///     equal; otherwise <see langword="false" />.
 	/// </summary>
-	bool AreConsideredEqual(T a, T b);
+	bool AreConsideredEqual(T actual, T expected);
 }
