+++ Source/aweXpect/Results/StringEqualityResult.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Text.RegularExpressions;
 using aweXpect.Core;
 using aweXpect.Options;
 
@@ -24,7 +25,7 @@ public class StringEqualityResult<TType, TThat>(
 ///     The result of an expectation with an underlying value of type <typeparamref name="TType" />.
 ///     <para />
 ///     In addition to the combinations from <see cref="AndOrResult{TType,TThat}" />, allows specifying
-///     options on the <see cref="StringMatcher" />.
+///     options on the <see cref="StringEqualityOptions" />.
 /// </summary>
 public class StringEqualityResult<TType, TThat, TSelf>(
 	ExpectationBuilder expectationBuilder,
@@ -33,15 +34,6 @@ public class StringEqualityResult<TType, TThat, TSelf>(
 	: AndOrResult<TType, TThat, TSelf>(expectationBuilder, returnValue)
 	where TSelf : StringEqualityResult<TType, TThat, TSelf>
 {
-	/// <summary>
-	///     Ignores casing when comparing the <see langword="string" />s.
-	/// </summary>
-	public TSelf IgnoringCase()
-	{
-		options.IgnoringCase();
-		return (TSelf)this;
-	}
-
 	/// <summary>
 	///     Ignores casing when comparing the <see langword="string" />s,
 	///     according to the <paramref name="ignoreCase" /> parameter.
@@ -73,6 +65,33 @@ public class StringEqualityResult<TType, TThat, TSelf>(
 		IEqualityComparer<string> comparer)
 	{
 		options.UsingComparer(comparer);
+		return (TSelf)this;
+	}
+	/// <summary>
+	///     Interprets the expected <see langword="string" /> as <see cref="Regex" /> pattern.
+	/// </summary>
+	public TSelf AsRegex()
+	{
+		options.AsRegex();
+		return (TSelf)this;
+	}
+
+	/// <summary>
+	///     Interprets the expected <see langword="string" /> as wildcard pattern.<br />
+	///     Supports * to match zero or more characters and ? to match exactly one character.
+	/// </summary>
+	public TSelf AsWildcard()
+	{
+		options.AsWildcard();
+		return (TSelf)this;
+	}
+
+	/// <summary>
+	///     Interprets the expected <see langword="string" /> to be exactly equal.
+	/// </summary>
+	public TSelf Exactly()
+	{
+		options.Exactly();
 		return (TSelf)this;
 	}
 }
