+++ Source/aweXpect/That/Delegates/ThatDelegateThrows.WithMessage.cs
@@ -8,12 +8,13 @@ public partial class ThatDelegateThrows<TException>
 	/// <summary>
 	///     Verifies that the thrown exception has a message equal to <paramref name="expected" />
 	/// </summary>
-	public StringMatcherResult<TException, ThatDelegateThrows<TException>>
-		WithMessage(
-			StringMatcher expected)
-		=> new(ExpectationBuilder.AddConstraint(it
+	public StringEqualityResult<TException, ThatDelegateThrows<TException>> WithMessage(string expected)
+	{
+		StringEqualityOptions? options = new();
+		return new StringEqualityResult<TException, ThatDelegateThrows<TException>>(ExpectationBuilder.AddConstraint(it
 				=> new ThatExceptionShould.HasMessageValueConstraint<TException>(
-					it, "with", expected)),
+					it, "with", expected, options)),
 			this,
-			expected);
+			options);
+	}
 }
