+++ Source/aweXpect/That/Exceptions/ThatExceptionShould.HaveMessage.cs
@@ -12,11 +12,14 @@ public partial class ThatExceptionShould<TException>
 	/// <summary>
 	///     Verifies that the actual exception has a message equal to <paramref name="expected" />
 	/// </summary>
-	public StringMatcherResult<TException?, ThatExceptionShould<TException>>
-		HaveMessage(StringMatcher expected)
-		=> new(ExpectationBuilder.AddConstraint(it
+	public StringEqualityResult<TException?, ThatExceptionShould<TException>>
+		HaveMessage(string expected)
+	{
+		var options = new StringEqualityOptions();
+		return new(ExpectationBuilder.AddConstraint(it
 				=> new ThatExceptionShould.HasMessageValueConstraint<TException>(
-					it, "have", expected)),
+					it, "have", expected, options)),
 			this,
-			expected);
+			options);
+	}
 }
